Refactor + Clean up + Add long poll logic when adding more tables
