+++ apps/studio/components/interfaces/Storage/ImportForeignSchemaDialog.tsx
@@ -36,11 +36,14 @@ export const ImportForeignSchemaDialog = ({
   onClose,
   circumstance = 'fresh',
 }: ImportForeignSchemaDialogProps) => {
-  const { data: project } = useSelectedProjectQuery()
   const { ref } = useParams()
+  const { data: project } = useSelectedProjectQuery()
   const [loading, setLoading] = useState(false)
   const [createSchemaSheetOpen, setCreateSchemaSheetOpen] = useState(false)
 
+  const { data: schemas } = useSchemasQuery({ projectRef: project?.ref! })
+
+  const { mutateAsync: createSchema } = useSchemaCreateMutation()
   const { mutateAsync: importForeignSchema } = useFDWImportForeignSchemaMutation({})
   const { mutateAsync: updateFDW } = useFDWUpdateMutation({
     onSuccess: () => {
@@ -48,8 +51,6 @@ export const ImportForeignSchemaDialog = ({
       onClose()
     },
   })
-
-  const { data: schemas } = useSchemasQuery({ projectRef: project?.ref! })
 
   const FormSchema = z.object({
     bucketName: z.string().trim(),
@@ -73,11 +74,9 @@ export const ImportForeignSchemaDialog = ({
     defaultValues: {
       bucketName,
       sourceNamespace: namespace,
-      targetSchema: '',
+      targetSchema: `fdw_analytics_${namespace}`,
     },
   })
-
-  const { mutateAsync: createSchema } = useSchemaCreateMutation()
 
   const onSubmit: SubmitHandler<z.infer<typeof FormSchema>> = async (values) => {
     const serverName = `${snakeCase(values.bucketName)}_fdw_server`
@@ -150,7 +149,7 @@ export const ImportForeignSchemaDialog = ({
       form.reset({
         bucketName,
         sourceNamespace: namespace,
-        targetSchema: '',
+        targetSchema: `fdw_analytics_${namespace}`,
       })
     }
   }, [visible, form, bucketName, namespace])
